services/account_service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-JWT-Extended==4.4.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.4
SQLAlchemy==1.3.24
psycopg2-binary==2.9.1
swagger-ui-bundle==0.0.9
Flask-Swagger-UI==4.11.1
pylint==3.0.3
pytest==6.2.5
python-dotenv==0.10.1
